Disable wayland client-side decorations
Enabling this macro makes the demo appear inside of a "window", with an exit
button and a border. However, this window is glitchy on some of our platforms
(mainly ccimx6, ccimx6qp and ccimx8mm) and blank in others (ccimx93). This is
probably related to how weston is configured depending on each platform's
graphics capabilities.

Disable these decorations so that the demo appears the same way in all
platforms.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>